Refactor : 중복 닉네임 조회 방식 변경

+++ src/main/java/io/urdego/urdego_user_service/domain/entity/User.java
@@ -71,7 +71,7 @@ public class User extends BaseTimeEntity{
 	@JoinColumn(name = "active_character_id",nullable = true)
 	private GameCharacter activeCharacter;
 
-	public static User create(UserSignUpRequest signUpRequest, int nicknameNum) {
+	public static User create(UserSignUpRequest signUpRequest, Long nicknameNum) {
 		return User.builder()
 				.name(signUpRequest.nickname())
 				.nickname(signUpRequest.nickname() +"#"+nicknameNum)
